requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
Faker==23.3.0
idna==3.6
phonenumbers==8.13.32
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==4.0.2
tzdata==2024.1
urllib3
auto-changelog
selenium==4.1.0
locust==2.5.1
chromedriver-autoinstaller==0.2.2
coverage
python-dotenv